requirements.txt
pydantic>=2.13.3
pydantic-core>=2.46.3
python-dotenv
langchain
langchain-core
langchain-community
langchain-chroma
langchain-groq
langchain-huggingface
langchain-text-splitters
pypdf
chromadb
sentence-transformers
flask
flask-cors
flask-sqlalchemy
flask-bcrypt
pyjwt
# Phase 7: Monitoring & Evaluation
ragas
langsmith
openai
datasets
